[HumanBenchmark][Backend][Users] refactor users endpoints
