+++ back-end/api/model-prisma.js
@@ -3,7 +3,6 @@ const { PrismaClient } = require("@prisma/client");
 
 const prisma = new PrismaClient();
 
-
 function calculateXP(movement) {
   // Should take user roughly 3 workouts to level up
   return (movement.reps * movement.sets * movement.weight) / 725;
@@ -37,17 +36,6 @@ function calculateOverallStat(bodyPartStats) {
 }
 
 module.exports = {
-  async findUsers(where) {
-    // GET http://localhost:5432/api/users?type=Recent
-    const users = await prisma.user.findMany({ where });
-    return users;
-  },
-
-  async findUserById(id) {
-    const user = await prisma.user.findUnique({ where: { id } });
-    return user;
-  },
-
   async findUsersLeaderboard(where) {
     const users = await prisma.user.findMany({
       where,
@@ -57,49 +45,6 @@ module.exports = {
       },
     });
     return users;
-  },
-
-  async createUser(data) {
-    const created = await prisma.user.create({ data });
-    return created;
-  },
-
-  async updateUser(id, data) {
-    const updated = await prisma.user.update({ where: { id }, data });
-    return updated;
-  },
-
-  async createBodyPartStat(userId, newBodyPartStat) {
-    const created = await prisma.bodyPartStat.create({ data: newBodyPartStat });
-
-    const bodyPartStats = await prisma.bodyPartStat.findMany({
-      where: { userId },
-    });
-
-    prisma.user.update({
-      where: { id: userId },
-      data: { bodyPartStats: { push: created } },
-    });
-    return created;
-  },
-
-  async createMuscleStat(userId, newMuscleStat) {
-    const createdMuscleStat = await prisma.muscleStat.create({
-      data: {
-        muscle: newMuscleStat.muscle,
-        bodyPart: newMuscleStat.bodyPart,
-        max: newMuscleStat.max,
-        user: { connect: { id: userId } },
-        exercise: newMuscleStat.exerciseId
-          ? { connect: { id: newMuscleStat.exerciseId } }
-          : undefined,
-        bodyPartStat: newMuscleStat.bodyPartStatId
-          ? { connect: { id: newMuscleStat.bodyPartStatId } }
-          : undefined,
-      },
-    });
-
-    return createdMuscleStat;
   },
 
   async getMuscleStats(userId) {
